'Refactored by Sourcery'

+++ update.py
@@ -44,7 +44,7 @@ def _updateCheckout(branch, update):
             f"https://github.com/Nuitka/Nuitka/archive/{branch}.zip", "nuitka.zip.tmp"
         )
 
-        with zipfile.ZipFile(f"nuitka.zip.tmp") as archive:
+        with zipfile.ZipFile('nuitka.zip.tmp') as archive:
             archive.extractall(".")
 
         os.unlink("nuitka.zip.tmp")
@@ -152,7 +152,7 @@ def updateDownloadPage():
             if v in ".-+":
                 parts.append(current)
                 current = ""
-            elif v.isdigit() and (current.isdigit() or current == ""):
+            elif v.isdigit() and (current.isdigit() or not current):
                 current += v
             elif v.isdigit():
                 parts.append(current)
@@ -268,7 +268,7 @@ def updateDownloadPage():
         return match.group(1)
 
     def makeRepositoryUrl(path):
-        return "https://nuitka.net/" + path
+        return f'https://nuitka.net/{path}'
 
     deb_info = {}
 
@@ -323,8 +323,7 @@ def updateDownloadPage():
 
         def splitVersion(v):
             for w in v.split("."):
-                for x in w.split("rc"):
-                    yield x
+                yield from w.split("rc")
 
         def compareVersion(v):
             v = v.split("-")
@@ -403,12 +402,11 @@ def updateDownloadPage():
 
         m1, m2, m3 = numberize(filename)[1:4]
 
-        if not m2:
-            result = "0.%d.%drc%d" % (m1, m3 / 100, (m3 / 10) % 10)
-        else:
-            result = "0.%d.%d.%d" % (m1, (m3 // 10), m3 % 10)
-
-        return result
+        return (
+            "0.%d.%drc%d" % (m1, m3 / 100, (m3 / 10) % 10)
+            if not m2
+            else "0.%d.%d.%d" % (m1, (m3 // 10), m3 % 10)
+        )
 
     findings = {
         "plain_prerelease": makePlain(max_pre_release),
@@ -718,7 +716,7 @@ compatible Python compiler,  `"download now" </doc/download.html>`_.\n""",
                 slug = "minor-" + slug.replace("nuitka-release", "release-nuitka")
 
             output_path = "doc/posts"
-            txt_path = os.path.join(output_path, slug + ".rst")
+            txt_path = os.path.join(output_path, f'{slug}.rst')
 
             if os.path.exists(txt_path):
                 pub_date = open(txt_path).readline().split(maxsplit=2)[2].strip()
@@ -807,8 +805,9 @@ def runPostProcessing():
 
     js_set_1 = ["jquery", "underscore", "doctools", "js/theme"]
     js_set_1_contents = "\n".join(
-        getFileContents("output/_static/" + js_name + ".js") for js_name in js_set_1
+        getFileContents(f'output/_static/{js_name}.js') for js_name in js_set_1
     )
+
 
     js_set_1_contents += """
 jQuery(function () {
@@ -822,7 +821,7 @@ jQuery(function () {
         js_set_1_contents
     )
 
-    putTextFileContents("output" + js_set_1_output_filename, js_set_1_contents)
+    putTextFileContents(f'output{js_set_1_output_filename}', js_set_1_contents)
 
     for filename in getFileList("output", only_suffixes=".html"):
         doc = html.fromstring(getFileContents(filename, mode="rb"))
@@ -838,7 +837,7 @@ jQuery(function () {
         css_links = doc.xpath("//link[@rel='stylesheet']")
         assert css_links
 
-        css_filenames = [
+        if css_filenames := [
             os.path.normpath(
                 "output/%s/%s"
                 % (
@@ -849,9 +848,7 @@ jQuery(function () {
             for css_link in css_links
             if "combined_" not in css_link.get("href")
             if "copybutton" not in css_link.get("href") or has_highlight
-        ]
-
-        if css_filenames:
+        ]:
             output_filename = "/_static/css/combined_%s.css" % getHashFromValues(
                 *css_filenames
             )
@@ -869,20 +866,18 @@ jQuery(function () {
                 merged_css = re.sub(r"/\*.*?\*/", "", merged_css, flags=re.S)
                 merged_css = re.sub(r"\s+\n", r"\n", merged_css, flags=re.M)
 
-                putTextFileContents(
-                    filename="output" + output_filename, contents=merged_css
-                )
+                putTextFileContents(filename=f'output{output_filename}', contents=merged_css)
 
             css_links[0].attrib["href"] = output_filename
             for css_link in css_links[1:]:
                 css_link.getparent().remove(css_link)
 
         for link in doc.xpath("//a[not(contains(@classes, 'intern'))]"):
-            if link.attrib["href"].startswith("http:") or link.attrib[
-                "href"
-            ].startswith("https:"):
-                if "nuitka.net" not in link.attrib["href"]:
-                    link.attrib["target"] = "_blank"
+            if (
+                link.attrib["href"].startswith("http:")
+                or link.attrib["href"].startswith("https:")
+            ) and "nuitka.net" not in link.attrib["href"]:
+                link.attrib["target"] = "_blank"
 
         (logo_img,) = doc.xpath("//img[@class='logo']")
 
@@ -1004,10 +999,9 @@ def runDeploymentCommand():
 
     branch = subprocess.check_output("git branch --show-current".split()).strip()
 
-    if branch == b"main":
-        if os.path.exists("output/robots.txt-operational"):
-            os.unlink("output/robots.txt")
-            os.rename("output/robots.txt.operational", "output/robots.txt")
+    if branch == b"main" and os.path.exists("output/robots.txt-operational"):
+        os.unlink("output/robots.txt")
+        os.rename("output/robots.txt.operational", "output/robots.txt")
 
     target_dir = "/var/www/" if branch == b"main" else "/var/www-staging/"
     command = (
